+++ packages/expo-barcode-scanner/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🎉 New features
 
 - Delete `prop-types` in favor of TypeScript. ([#8678](https://github.com/expo/expo/pull/8678) by [@EvanBacon](https://github.com/EvanBacon))
+- `BarCodeScanner` is now returning barcode's bounding box on iOS. ([#8865](https://github.com/expo/expo/pull/8865) by [@lukmccall](https://github.com/lukmccall))
 
 ### 🐛 Bug fixes